Collect sys_getframe tests into a single file
